c++-mode: clang-format before save.
This also tries to find the newest clang-format on the system. On my
current primary project, there's stuff in its .clang-format that
clang-format-17 (which is installed) understands but clang-format-14
(which is also installed) does not. Since /usr/bin/clang-format points
to clang-format-14, this results in annoyance on save.

Fixing it here is a lot easier than trying to make my Linux distro
point /usr/bin/clang-format elsewhere.