Add snoopy gopackagesdriver
When gopls complains that its gopackagesdriver (sometimes spelled
"go/packages driver") can't resolve something, that's difficult to
debug.

This commit adds a snoopy gopackagesdriver, which is a simple Python
script that invokes the real driver and logs its stdio streams to
files in /tmp. While the output is quite verbose, it can be helpful to
figure out why a particular package is not resolvable.

It appears that stdin and stdout each contain one JSON object, while
stderr is the usual Bazel spew.

To toggle snoopiness, edit gopls-wrapper to set PKGDRIVER to the one
you want.